requirements.txt
opencv-python
pytesseract
adb-shell
numpy~=1.26.4
torch~=2.4.0
pillow~=10.3.0
torchvision~=0.19.0
tesseract